fix: merge issues hubadminviewcontroller
