+++ labs-ios-starter/View Controllers/HubAdmin/HomeScreen/HubAdminViewController.swift
@@ -8,10 +8,6 @@
 
 import UIKit
 
-//protocol DashboardLayoutDelegate: AnyObject {
-//    func collectionView(_ collectionView: UICollectionView, heightForCellAtIndexPath indexPath: IndexPath) -> CGFloat
-//}
-
 class HubAdminViewController: UIViewController {
 
     // MARK: - IBOutlets
@@ -20,31 +16,10 @@ class HubAdminViewController: UIViewController {
 
     // MARK: - Properties
     private let controller = BackendController.shared
-    
     private var hubs: [Hub] = []
     private var selectedHub: Hub?
-    
-   
-    
-    
-    var productionReportArray: [HubDailyProduction] = [HubDailyProduction(dictionary: ["id" : "5",
-                                                                               "barsProduced": 10,
-                                                                               "soapmakersWorked": 3,
-                                                                               "soapmakerHours": 17,
-                                                                               "date":  "2020-10-15"
-    ])!,
-    HubDailyProduction(dictionary: ["id" : "5",
-                                                                               "barsProduced": 10,
-                                                                               "soapmakersWorked": 3,
-                                                                               "soapmakerHours": 17,
-                                                                               "date":  "2020-10-10"
-    ])!]
-
-
-
-    // MARK: - Properties
-    weak var delegate: DashboardLayoutDelegate?
-
+    private var productionReportArray: [HubDailyProduction] = []
+    private var impactStats: ImpactStats?
     private var numberOfColumns = 2
     private let cellPadding: CGFloat = 8
     private var cache: [UICollectionViewLayoutAttributes] = []
@@ -58,20 +33,20 @@ class HubAdminViewController: UIViewController {
     }
 
     // MARK: - View Lifecycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        fetchAll()
-      
-        tableView.delegate = self
-
+    override func viewWillAppear(_ animated: Bool) {
         guard let userhub = controller.loggedInUser.hub else {return}
         controller.hubs[userhub.id] = userhub
-
+        fetchAll()
         updateViewsProductionReports()
-
         updateViewsImpactStats()
     }
 
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        tableView.delegate = self
+    }
+
+    //MARK: - Methods
     private func fetchAll() {
         controller.initialFetch(userId: controller.loggedInUser.id) { (error) in
             if let error = error {
@@ -82,117 +57,131 @@ class HubAdminViewController: UIViewController {
                 print(self.controller.loggedInUser.id)
             }
             print("\(self.controller.users)")
-            print("\(self.controller.properties)")
             print("\(self.controller.pickups)")
-            print("\(self.controller.payments)")
             print("\(self.controller.productionReports)")
             print("\(self.controller.hubs)")
-            print("\(self.controller.pickupCartons)")
-            print("\(self.controller.hospitalityContracts)")
         }
     }
-    
 
     private func updateViewsImpactStats() {
         guard let selectedHub = controller.loggedInUser.hub?.id else { return }
-        controller.impactStatsByHubId(id: selectedHub) { (error) in
+        controller.impactStatsByHubId(id: selectedHub) { (data, error) in
             if let error = error {
                 print("Error fetching stats \(error)")
-                
+                return
             }
+
+            guard let data = self.controller.loggedInUser.hub?.impact else { return }
+
+            self.impactStats = data
+            self.overallBreakDown()
+
             DispatchQueue.main.async {
                 self.collectionView.reloadData()
             }
         }
     }
-    
 
     private func updateViewsProductionReports() {
         guard let selectedHub = controller.loggedInUser.hub else { return }
         controller.productionReportsByHubId(hubId: selectedHub.id) { (error) in
             if let error = error {
-                print("Error fetching stats \(error)")
+                print("Error fetching production reports \(error)")
+                return
+            }
+            for productionReport in self.controller.productionReports.values {
+                self.productionReportArray.append(productionReport)
             }
 
             DispatchQueue.main.async {
-                self.collectionView.reloadData()
+                self.tableView.reloadData()
             }
         }
     }
 
     private func overallBreakDown() -> (Int, Int, Int, Int) {
-        guard let selectedHub = selectedHub else { return (0,0,0,0) }
-        let bottlesRecycled = selectedHub.impact?.bottlesRecycled ?? 0
-        let soapRecycled = selectedHub.impact?.soapRecycled ?? 0
-        let paperRecycled = selectedHub.impact?.paperRecycled ?? 0
-        let linensRecycled = selectedHub.impact?.linensRecycled ?? 0
+        let bottlesRecycled = impactStats?.bottlesRecycled ?? 1
+        let soapRecycled = impactStats?.soapRecycled ?? 1
+        let paperRecycled = impactStats?.paperRecycled ?? 1
+        let linensRecycled = impactStats?.linensRecycled ?? 1
 
         let total = bottlesRecycled + soapRecycled + paperRecycled + linensRecycled
         let soapPercentage = (Double(soapRecycled) / Double(total)) * 100
         let bottlesPercentage = (Double(bottlesRecycled) / Double(total)) * 100
         let paperPercentage = (Double(paperRecycled) / Double(total)) * 100
         let linensPercentage = (Double(linensRecycled) / Double(total)) * 100
-        return (Int(soapPercentage), Int(linensPercentage), Int(bottlesPercentage), Int(paperPercentage))
 
+        return (Int(soapPercentage), Int(linensPercentage), Int(bottlesPercentage), Int(paperPercentage))
+    }
+
+    // MARK: - Navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        switch segue.identifier {
+        case "HubAdminEditProdRepSegue":
+            if let indexPath = tableView.indexPathForSelectedRow?.row {
+                guard let editProdReportVC = segue.destination as? HubAdminEditProductionReportViewController else {return}
+
+                editProdReportVC.productionReport = productionReportArray[indexPath]
+            }
+        case "HubAdminNewProdRepSegue":
+            return
+
+        case "SettingsSegue":
+            return
+
+        default: preconditionFailure("unexpected segue identifier")
+        }
     }
 }
-    // MARK: - Navigation
-
+// MARK: - UICollectionView Extension
 extension HubAdminViewController: UICollectionViewDelegateFlowLayout {
-    
+
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
         let width : CGFloat
         let height : CGFloat
-        
-       
-            width = collectionView.bounds.width - 40
-            height = 200
-       
-     
+
+        width = collectionView.bounds.width - 40
+        height = 200
+
         return CGSize(width: width, height: height)
     }
-    
+
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
         return UIEdgeInsets(top: 20.0, left: 20, bottom: 20, right: 20)
     }
-
 }
-
 
 extension HubAdminViewController: UICollectionViewDelegate, UICollectionViewDataSource {
 
-    // MARK: UICollectionViewDataSource
     func numberOfSections(in collectionView: UICollectionView) -> Int {
         return 1
     }
-    
+
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        
         return 1
     }
 
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-            guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "HubAdminImpactStatisticsOverallCell", for: indexPath) as? HubAdminCollectionViewCell else { return UICollectionViewCell() }
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "HubAdminImpactStatisticsOverallCell", for: indexPath) as? HubAdminCollectionViewCell else { return UICollectionViewCell() }
 
-            cell.statsTuple = overallBreakDown()
-            
-            return cell
-        }
+        cell.statsTuple = overallBreakDown()
+
+        return cell
+    }
 }
 
+// MARK: - UITableView Extension
 extension HubAdminViewController: UITableViewDataSource,  UITableViewDelegate {
-    
+
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return productionReportArray.count
     }
-    
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         guard let cell = tableView.dequeueReusableCell(withIdentifier: "HubAdminProductionReportsCell", for: indexPath) as? HubAdminMainTableViewCell else {fatalError("oops")}
         let report = productionReportArray[indexPath.row]
         cell.productionReports = report
-     
+
         return cell
     }
-
 }
